MLOps and MLSD/Practice/week1-ai-ml-system-diagram/ML-System-Pipeline/requirements.txt
scikit-learn==1.2.2
fastapi==0.85.0
uvicorn==0.18.3
prometheus-client==0.16.0
mlflow==2.2.0
joblib==1.2.0
pandas==1.5.3
numpy==1.23.5
pytest==7.3.1
PyYAML==6.0
